alguns detalhes componente mej #14
