+++ src/app/sobre-nos/page.js
@@ -3,7 +3,6 @@ import AboutFaisca from "@components/sobreNos/AboutFaisca"
 import MembersArea from "@components/sobreNos/MembersArea"
 import Structure from "@/components/sobreNos/Structure"
 import MEJ from "@components/sobreNos/MEJ"
-import MEJO from "@components/sobreNos/MEJO"
 
 import styles from "@styles/pages.module.css"
 
@@ -14,7 +13,7 @@ function SobreNos(){
             <AboutFaisca />
             <MembersArea />
             <Structure/> 
-            <MEJO/>
+            <MEJ/>
         </div>
     )
 }
