+++ src/components/home/HomeCarrousel/index.js
@@ -1,4 +1,5 @@
 'use client'
+
 import React from 'react';
 import 'owl.carousel/dist/assets/owl.carousel.css';
 import 'owl.carousel/dist/assets/owl.theme.default.css';
