+++ src/components/sobreNos/MEJ/index.js
@@ -1,40 +1,29 @@
-import styles from "./mej.module.css";
-import brasil from "@img/sobreNos/brasil.png"
-import NejPoa from "@img/sobreNos/NejPoa.png"
-import fejers from "@img/sobreNos/fejers.png"
-import sla from "@img/sobreNos/sla.png"
-import img from "@img/sobreNos/MEJ/MEJ.jpg"
+import styles from './mej.module.css'
+import Image from "next/image"
+import rectagle from '@img/sobreNos/MEJ/teste.png'
+import brasilJR from "@img/sobreNos/MEJ/brasil.png"
+import NejPoa from "@img/sobreNos/MEJ/NejPoa.png"
+import fejers from "@img/sobreNos/MEJ/fejers.png"
+import sla from "@img/sobreNos/MEJ/sla.png"
 
-import Image from "next/image";
-
-function MEJ ()
-{
+function MEJO() {
     return (
         <div className={styles.container}>
-            <Image className={styles.imagem} src={img} width={335} height={381} alt="imagem"/>
-            <div className={styles.bar}/>
-            <div className={styles.containerParagrafo}>
-                <h1 className={styles.titulo}>
-                    MEJ
-                </h1>
-                <p className={styles.paragrafo}>
-                    Nossa missão é formar, por meio da vivência empresarial, 
-                    lideranças comprometidas e capazes de transformar o país em 
-                    um brasil empreendedor.
-                    Desde 2010, o MEJ já impactou mais de R$ 70.000.000,00 
-                    na economia brasileira, que são integralmente reinvestidos 
-                    na educação empreendedora dos estudantes.
-                </p>
-                
-                <div className={styles.conjuntoImagens}>
-                    <Image className={styles.logos} src={brasil} width={99} height={88} alt="brasil" />
-                    <Image className={styles.logos} src={NejPoa} width={82} height={82} alt="NejPoa" />
-                    <Image className={styles.logos} src={fejers} width={142} height={55} alt="Fejers" />
-                    <Image className={styles.logos} src={sla} width={76} height={90} alt="SLA" />
+            <Image src={rectagle} alt={"error"} width={0} height={420} />
+            <div className={styles.bar}></div>
+            <div className={styles.rightContent}>
+                <h1>Movimento Empresa Jr</h1>
+                <p>Nossa missão é formar, por meio da vivência empresarial, lideranças comprometidas e capazes de transformar o país em um brasil empreendedor. Desde 2010,
+                    o MEJ já impactou mais de R$ 70.000.000,00 na economia brasileira, que são integralmente reinvestidos na educação empreendedora dos estudantes.</p>
+                <div className={styles.apoiadores}>
+                    <Image className={styles.logos} src={brasilJR} width={0} height={80} alt="brasil" />
+                    <Image className={styles.logos} src={NejPoa} width={0} height={75} alt="NejPoa" />
+                    <Image className={styles.logos} src={fejers} width={0} height={47} alt="Fejers" />
+                    <Image className={styles.logos} src={sla} width={0} height={82} alt="SLA" />
                 </div>
-                
             </div>
         </div>
     )
 }
-export default MEJ
+
+export default MEJO;
