+++ src/components/sobreNos/MEJ/mej.module.css
@@ -1,86 +1,89 @@
-.container{
+.container {
     display: flex;
-    align-items: center;
     flex-direction: row;
-    flex-wrap: nowrap;
-    justify-content: center;
-    padding: 5%;
-    margin-top: 12vh;
-    margin-bottom: 15vh;
-}
-
-.containerParagrafo{
     color: white;
+    width: 75%;
+    margin-left: auto;
+    margin-right: auto;
+    margin-top: 15vh;
+    margin-bottom: 25vh;
+    align-items: center;
 }
-.paragrafo{
-    font-size: 1vw;
-    margin: 0px;
-    text-align: left;
-    margin-bottom: 2em;
-    margin-right: 12vh;
- }
 
- .titulo{
-    font-size: 2.7vw;
-    font-weight: 500;
-    text-align: left;
-    margin-bottom: 1.7vh;
- }
-
- .conjuntoImagens{
-   
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    width: 50%;
- }
-
- .logos{
-    height: 20%;
-    width: 20%;
- }
-
- .bar{
-    height: 68vh;
+.bar {
+    height: 81vh;
     border-left: 1px solid;
     color: white;
     margin: 0px 5em;
 }
 
- @media (max-width: 767px) {
+.rightContent h1 {
+    font-size: 2.7vw;
+    font-weight: 500;
+    margin-bottom: 2.3vh;
+}
+
+.rightContent p {
+    font-size: 1vw;
+    font-weight: 300;
+    line-height: normal;
+}
+
+.apoiadores {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    margin-top: 4.3vh;
+    width: 75%;
+}
+
+.imge {
+    width: 100%; 
+    height: auto; 
+}
+
+.logos {
+    width: auto;
+    height: 4.5vw; 
+}
+
+@media (max-width: 576px) {
     .container {
-      flex-direction: column;
+        flex-direction: column;
+        align-items: center;
     }
-    .bar{
+
+    .bar {
         display: none;
     }
-    .containerParagrafo{
-        color: white;
-        display: flex;
-        flex-direction: column;
-        width: 80%;
-        flex-wrap: nowrap;
-        align-content: left;
-        justify-content: left;
-        align-items: left;
+
+    .rightContent h1 {
+        font-size: 7.52vw;
+        margin-top: 3.3vh;
+        width: 90%;
+        margin-left: auto;
+        margin-right: auto;
     }
-    .imagem{
-        margin-bottom: 4em;
-        width: 80%;
+    
+    .rightContent p {
+        font-size: 2.82vw;
+        width: 90%;
+        margin-left: auto;
+        margin-right: auto;
     }
-    .conjuntoImagens{
-        margin-top: 10px;
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        margin-right: 2em;
-        width: 100%;
+
+    .imge {
+        width: 80%; 
+        height: auto; 
     }
-    .paragrafo{
-        margin-bottom: 2em;
-        font-size: 2.84vw;
+
+    .logos {
+        height: 9vw; 
     }
-    .titulo{
-        font-size: 7.58vw;
+
+    .apoiadores{
+        width: 85%;
+        margin-left: auto;
+        margin-right: auto; 
     }
-  }
+}
