--- src/components/sobreNos/MEJO/mejo.module.css
@@ -1,89 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: row;
-    color: white;
-    width: 75%;
-    margin-left: auto;
-    margin-right: auto;
-    margin-top: 15vh;
-    margin-bottom: 25vh;
-    align-items: center;
-}
-
-.bar {
-    height: 81vh;
-    border-left: 1px solid;
-    color: white;
-    margin: 0px 5em;
-}
-
-.rightContent h1 {
-    font-size: 2.7vw;
-    font-weight: 500;
-    margin-bottom: 2.3vh;
-}
-
-.rightContent p {
-    font-size: 1vw;
-    font-weight: 300;
-    line-height: normal;
-}
-
-.apoiadores {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    margin-top: 4.3vh;
-    width: 75%;
-}
-
-.imge {
-    width: 100%; 
-    height: auto; 
-}
-
-.logos {
-    width: auto;
-    height: 4.5vw; 
-}
-
-@media (max-width: 576px) {
-    .container {
-        flex-direction: column;
-        align-items: center;
-    }
-
-    .bar {
-        display: none;
-    }
-
-    .rightContent h1 {
-        font-size: 7.52vw;
-        margin-top: 3.3vh;
-        width: 90%;
-        margin-left: auto;
-        margin-right: auto;
-    }
-    
-    .rightContent p {
-        font-size: 2.82vw;
-        width: 90%;
-        margin-left: auto;
-        margin-right: auto;
-    }
-
-    .imge {
-        width: 80%; 
-        height: auto; 
-    }
-
-    .logos {
-        height: 9vw; 
-    }
-
-    .apoiadores{
-        width: 85%;
-        margin-left: auto;
-        margin-right: auto; 
-    }
-}
